requirements.txt
numpy
opencv-python
pyserial
simpleaudio

